+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Referernce type signature parsing
 - Tweaks to printing/formatting of help messages in diagnostics
 - Tweaks/updates to links on website
+- Fix bug with import declaration parsing not accepting comments in certain places
 
 ## 0.10.1
 - Fix bug in release workflow that prevented wright binaries from being built